fix: don't use `toolbox` as constant for FQDN
issue: https://github.com/containers/toolbox/issues/969

Setting the hostname to toolbox causes timeouts whenever anything tries
to resolve the name of the machine - for example `sudo` does this.

This change makes it so the FQDN is set to
`${container_name}.${hostname}` as recommended in the linked issue.

After this change commands can properly resolve the local FQDN.

Signed-off-by: Alex Diaz <alex@akdev.xyz>