feat: improve wrapper script
turns out doing toolbox in podman is a bit difficult.

toolbox passes itself as the init system of the toolboxes so having that
be containerized is a bit difficult (though it seems possible)

pulling the image and then copying the binaries to the host is a nice
workaround - it enables quick and easy deployment and doesn't need root.